프레임 부분 수정 및 빌드 최신화

+++ 011. 인덱스 버퍼/source/base/main.cpp
@@ -134,7 +134,7 @@ HRESULT CALLBACK OnInit()
 	// 투영행렬을 설정합니다.
 	D3DXMATRIXA16 matProjection;
 	D3DXMatrixPerspectiveFovLH(&matProjection, (D3DX_PI / 4.0f),
-		(g_pMainDX->getClientWidth() / g_pMainDX->getClientHeight()),
+		static_cast<FLOAT>((g_pMainDX->getClientWidth() / g_pMainDX->getClientHeight())),
 		1.0f, 1000.0f);
 	g_pD3DDevice9->SetTransform(D3DTS_PROJECTION, &matProjection);
